minimize meta module
