--- packages/libraries/cli/src/helpers/errors/ClientError.ts
@@ -1,13 +1,0 @@
-import { GraphQLError } from 'graphql';
-
-export class ClientError extends Error {
-  constructor(
-    message: string,
-    public response: {
-      errors?: readonly GraphQLError[];
-      headers: Headers;
-    },
-  ) {
-    super(message);
-  }
-}
