--- packages/libraries/cli/src/helpers/errors/_.ts
@@ -1,12 +1,0 @@
-export { CommandError } from '@oclif/core/lib/interfaces';
-
-export {
-  CLIError,
-  RequiredArgsError,
-  FailedFlagValidationError,
-  ArgInvalidOptionError,
-} from '@oclif/core/lib/parser/errors';
-
-export * from './Failure';
-
-export * from './ClientError';
