--- packages/libraries/cli/src/helpers/errors/__.ts
@@ -1,1 +1,0 @@
-export * as Errors from './_';
