--- packages/libraries/cli/src/helpers/typebox/__.ts
@@ -1,1 +1,0 @@
-export * as T from './_';
