+++ packages/libraries/cli/src/helpers/typebox/_namespace.ts
@@ -1,0 +1,1 @@
+export * as T from './_exports';
