--- packages/libraries/cli/src/helpers/typebox/value/__.ts
@@ -1,1 +1,0 @@
-export * as Value from './_';
