--- packages/libraries/cli/src/output/__.ts
@@ -1,1 +1,0 @@
-export * as Output from './_';
