+++ packages/libraries/cli/src/output/_namespace.ts
@@ -1,0 +1,1 @@
+export * as Output from './_exports';
