--- packages/libraries/cli/src/output/result/_.ts
@@ -1,2 +1,0 @@
-export * from './failure';
-export * from './success';
