--- packages/libraries/cli/src/output/result/__.ts
@@ -1,5 +1,0 @@
-import { Failure } from './failure';
-import { Success } from './success';
-
-export type Result = Failure | Success;
-export * as Result from './_';
